box: move Lua tuple format wrapper to new `tuple_format` submodule
Since we are going to make public Lua tuple format objects, introduce a new
`tuple_format` submodule and move the code related to these objects there.

Needed for #4693

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring
