+++ src/box/lua/init.c
@@ -48,6 +48,7 @@
 
 #include "box/lua/error.h"
 #include "box/lua/tuple.h"
+#include "box/lua/tuple_format.h"
 #include "box/lua/call.h"
 #include "box/lua/slab.h"
 #include "box/lua/index.h"
@@ -743,6 +744,7 @@ box_lua_init(struct lua_State *L)
 	lua_pop(L, 1);
 
 	box_lua_error_init(L);
+	box_lua_tuple_format_init(L);
 	box_lua_tuple_init(L);
 	box_lua_call_init(L);
 	box_lua_cfg_init(L);
