+++ src/box/lua/misc.cc
@@ -44,7 +44,6 @@
 #include "box/read_view.h"
 #include "box/space_cache.h"
 #include "box/tuple.h"
-#include "box/tuple_format.h"
 #include "box/txn.h"
 #include "box/xrow.h"
 #include "core/diag.h"
@@ -55,8 +54,6 @@
 #include "mpstream/mpstream.h"
 #include "small/region.h"
 #include "small/rlist.h"
-
-static uint32_t CTID_STRUCT_TUPLE_FORMAT_PTR;
 
 /** {{{ Miscellaneous utils **/
 
@@ -360,90 +357,6 @@ lbox_txn_set_isolation(struct lua_State *L)
 	return 1;
 }
 
-/** {{{ Utils to work with tuple_format. **/
-
-struct tuple_format *
-lbox_check_tuple_format(struct lua_State *L, int narg)
-{
-	uint32_t ctypeid;
-	struct tuple_format *format =
-		*(struct tuple_format **)luaL_checkcdata(L, narg, &ctypeid);
-	if (ctypeid != CTID_STRUCT_TUPLE_FORMAT_PTR) {
-		luaL_error(L, "Invalid argument: 'struct tuple_format *' "
-			   "expected, got %s)",
-			   lua_typename(L, lua_type(L, narg)));
-	}
-	return format;
-}
-
-static int
-lbox_tuple_format_gc(struct lua_State *L)
-{
-	struct tuple_format *format =  lbox_check_tuple_format(L, 1);
-	tuple_format_unref(format);
-	return 0;
-}
-
-static int
-lbox_push_tuple_format(struct lua_State *L, struct tuple_format *format)
-{
-	struct tuple_format **ptr = (struct tuple_format **)
-		luaL_pushcdata(L, CTID_STRUCT_TUPLE_FORMAT_PTR);
-	*ptr = format;
-	tuple_format_ref(format);
-	lua_pushcfunction(L, lbox_tuple_format_gc);
-	luaL_setcdatagc(L, -2);
-	return 1;
-}
-
-static int
-lbox_tuple_format_new(struct lua_State *L)
-{
-	assert(CTID_STRUCT_TUPLE_FORMAT_PTR != 0);
-	int top = lua_gettop(L);
-	if (top == 0)
-		return lbox_push_tuple_format(L, tuple_format_runtime);
-	assert(top == 1 && lua_istable(L, 1));
-	uint32_t count = lua_objlen(L, 1);
-	if (count == 0)
-		return lbox_push_tuple_format(L, tuple_format_runtime);
-	struct region *region = &fiber()->gc;
-	size_t region_svp = region_used(region);
-	struct field_def *fields = xregion_alloc_array(region,
-						       struct field_def, count);
-	for (uint32_t i = 0; i < count; ++i) {
-		size_t len;
-		fields[i] = field_def_default;
-		lua_pushinteger(L, i + 1);
-		lua_gettable(L, 1);
-		lua_pushstring(L, "name");
-		lua_gettable(L, -2);
-		assert(! lua_isnil(L, -1));
-		const char *name = lua_tolstring(L, -1, &len);
-		fields[i].name = (char *)xregion_alloc(region, len + 1);
-		memcpy(fields[i].name, name, len);
-		fields[i].name[len] = '\0';
-		lua_pop(L, 1);
-		lua_pop(L, 1);
-	}
-	struct tuple_dictionary *dict = tuple_dictionary_new(fields, count);
-	region_truncate(region, region_svp);
-	if (dict == NULL)
-		return luaT_error(L);
-	struct tuple_format *format = runtime_tuple_format_new(dict);
-	/*
-	 * Since dictionary reference counter is 1 from the
-	 * beginning and after creation of the tuple_format
-	 * increases by one, we must decrease it once.
-	 */
-	tuple_dictionary_unref(dict);
-	if (format == NULL)
-		return luaT_error(L);
-	return lbox_push_tuple_format(L, format);
-}
-
-/* }}} */
-
 /** {{{ Read view utils. **/
 
 void
@@ -512,7 +425,6 @@ box_lua_misc_init(struct lua_State *L)
 	static const struct luaL_Reg boxlib_internal[] = {
 		{"prepare_auth", lbox_prepare_auth},
 		{"select", lbox_select},
-		{"new_tuple_format", lbox_tuple_format_new},
 		{"txn_set_isolation", lbox_txn_set_isolation},
 		{"read_view_list", lbox_read_view_list},
 		{"read_view_status", lbox_read_view_status},
@@ -523,10 +435,4 @@ box_lua_misc_init(struct lua_State *L)
 	luaL_findtable(L, LUA_GLOBALSINDEX, "box.internal", 0);
 	luaL_setfuncs(L, boxlib_internal, 0);
 	lua_pop(L, 1);
-
-	int rc = luaL_cdef(L, "struct tuple_format;");
-	assert(rc == 0);
-	(void) rc;
-	CTID_STRUCT_TUPLE_FORMAT_PTR = luaL_ctypeid(L, "struct tuple_format *");
-	assert(CTID_STRUCT_TUPLE_FORMAT_PTR != 0);
 }
