+++ test/box/misc.result
@@ -996,7 +996,7 @@ s:drop()
 -- no errors.
 --
 -- Without argument it is equivalent to new_tuple_format({})
-tuple_format = box.internal.new_tuple_format()
+tuple_format = box.internal.tuple_format.new()
 ---
 ...
 -- If no type that type == "any":
@@ -1009,18 +1009,18 @@ format[1] = {}
 format[1].name = 'aaa'
 ---
 ...
-tuple_format = box.internal.new_tuple_format(format)
+tuple_format = box.internal.tuple_format.new(format)
 ---
 ...
 -- Function space:format() without arguments returns valid format:
-tuple_format = box.internal.new_tuple_format(box.space._space:format())
+tuple_format = box.internal.tuple_format.new(box.space._space:format())
 ---
 ...
 -- Check is_nullable option fo field
 format[1].is_nullable = true
 ---
 ...
-tuple_format = box.internal.new_tuple_format(format)
+tuple_format = box.internal.tuple_format.new(format)
 ---
 ...
 --
